+++ pyproject.toml
@@ -1,0 +1,41 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "stamp"
+version = "0.0.1"
+authors = [
+  { name="Example Author", email="author@example.com" },
+]
+description = "A protocol for Solid Tumor Associative Modeling in Pathology"
+readme = "README.md"
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+  "pandas~=2.1",
+  "matplotlib~=3.8",
+  "scikit-learn~=1.3",
+  "tqdm~=4.66",
+  "fastai~=2.7",
+  "torch~=2.0",
+  "torchvision~=0.15",
+  "h5py~=3.10",
+  "einops~=0.7",
+  "omegaconf~=2.3",
+  "openslide-python~=1.3",
+  "opencv-python~=4.8",
+  "numba~=0.58",
+  "gdown~=4.7"
+]
+
+[project.scripts]
+"stamp" = "stamp.cli:main"
+
+[project.urls]
+"Homepage" = "https://github.com/Avic3nna/STAMP"
+"Bug Tracker" = "https://github.com/Avic3nna/STAMP/issues"
