--- requirements.txt
@@ -1,14 +1,0 @@
-pandas==2.1.1
-matplotlib==3.8.0
-scikit-learn==1.3.1
-tqdm==4.66.1
-fastai==2.7.12
-torch==2.0.1
-torchvision==0.15.2
-h5py==3.10.0
-einops==0.7.0
-omegaconf==2.3.0
-openslide-python==1.3.1
-opencv-python==4.8.1.78
-numba==0.58.0
-gdown==4.7.1
